move code caching to new module
