+++ xonsh/tools.py
@@ -21,8 +21,6 @@ import os
 import re
 import sys
 import ctypes
-import hashlib
-import marshal
 import builtins
 import platform
 import traceback
@@ -1133,120 +1131,3 @@ class CommandsCache(Set):
         allcmds |= set(builtins.aliases.keys())
         self._cmds_cache = frozenset(allcmds)
         return self._cmds_cache
-
-
-def _splitpath(path, sofar=[]):
-    folder, path = os.path.split(path)
-    if path == "":
-        return sofar[::-1]
-    elif folder == "":
-        return (sofar + [path])[::-1]
-    else:
-        return _splitpath(folder, sofar + [path])
-
-_CHARACTER_MAP = {chr(o): '_%s' % chr(o+32) for o in range(65, 91)}
-_CHARACTER_MAP.update({'.': '_.', '_': '__'})
-
-
-def _cache_renamer(path):
-    path = os.path.abspath(path)
-    o = [''.join(_CHARACTER_MAP.get(i, i) for i in w) for w in _splitpath(path)]
-    o[-1] = "{}.{}".format(o[-1], sys.implementation.cache_tag)
-    return o
-
-
-def _make_if_not_exists(dirname):
-    if not os.path.isdir(dirname):
-        os.makedirs(dirname)
-
-def _should_use_cache(execer, mode):
-    if mode == 'exec':
-        return ((execer.scriptcache or
-                    execer.cacheall) and
-                (builtins.__xonsh_env__['XONSH_CACHE_SCRIPTS'] or
-                    builtins.__xonsh_env__['XONSH_CACHE_EVERYTHING']))
-    else:
-        return (execer.cacheall or builtins.__xonsh_env__['XONSH_CACHE_EVERYTHING'])
-
-
-def run_script_with_cache(filename, execer, glb=None, loc=None, mode='exec'):
-    """
-    Run a script, using a cached version if it exists (and the source has not
-    changed), and updating the cache as necessary.
-    """
-    run_cached = False
-    use_cache = _should_use_cache(execer, mode)
-    if use_cache:
-        datadir = builtins.__xonsh_env__['XONSH_DATA_DIR']
-        cachedir = os.path.join(datadir, 'xonsh_script_cache')
-        cachefname = os.path.join(cachedir, *_cache_renamer(filename))
-        if os.path.isfile(cachefname):
-            if os.stat(cachefname).st_mtime >= os.stat(filename).st_mtime:
-                with open(cachefname, 'rb') as cfile:
-                    ccode = marshal.load(cfile)
-                    run_cached = True
-    if not run_cached:
-        with open(filename, 'r') as f:
-            code = f.read()
-        if not code.endswith('\n'):
-            code += '\n'
-        try:
-            old_filename = execer.filename
-            execer.filename = filename
-            ccode = execer.compile(code, glbs=glb, locs=loc, mode=mode)
-            if use_cache:
-                _make_if_not_exists(os.path.dirname(cachefname))
-                with open(cachefname, 'wb') as cfile:
-                    marshal.dump(ccode, cfile)
-        except:
-            raise
-        finally:
-            execer.filename = old_filename
-    if mode in {'exec', 'single'}:
-        func = exec
-    else:
-        func = eval
-    func(ccode, glb, loc)
-
-
-def run_code_with_cache(code, execer, glb=None, loc=None, mode='exec'):
-    """
-    Run a piece of code, using a cached version if it exists, and updating the
-    cache as necessary.
-    """
-    run_cached = False
-    use_cache = _should_use_cache(execer, mode)
-    if isinstance(code, str):
-        _code = code.encode()
-    else:
-        _code = code
-    filename = "{}.{}".format(hashlib.md5(_code).hexdigest(),
-                              sys.implementation.cache_tag)
-    if use_cache:
-        datadir = builtins.__xonsh_env__['XONSH_DATA_DIR']
-        cachedir = os.path.join(datadir, 'xonsh_code_cache')
-        cachefname = os.path.join(cachedir, filename)
-        if os.path.isfile(cachefname):
-            with open(cachefname, 'rb') as cfile:
-                ccode = marshal.load(cfile)
-                run_cached = True
-    if not run_cached:
-        if not code.endswith('\n'):
-            code += '\n'
-        try:
-            old_filename = execer.filename
-            execer.filename = filename
-            ccode = execer.compile(code, glbs=glb, locs=loc, mode=mode)
-            if use_cache:
-                _make_if_not_exists(cachedir)
-                with open(cachefname, 'wb') as cfile:
-                    marshal.dump(ccode, cfile)
-        except:
-            raise
-        finally:
-            execer.filename = old_filename
-    if mode in {'exec', 'single'}:
-        func= exec
-    else:
-        func = eval
-    func(ccode, glb, loc)
